ref: Manually measure various performance costs behind issue list
This is a terribly kludgey and improper form of performance measurement,
if it can even be called that. It is a temporary means of exploratory
debugging that I am sharing for the sake of discussion.

Do not merge this.

(The Stopwatch class may be worth keeping if others are interested in
doing a similar kind of exploration. Even so, I would lean against
merging it without any permanent code that calls it.)