perf(snuba): Avoid selecting IDs when the result set is too large
The changed query was used both to determine the size of the group and
to obtain a set of IDs within the group for later use. The IDs are not
used if the size places us in the "too many candidates" case.

Performance data suggest that the time cost of selecting IDs is
prohibitively high when there are many rows in the table, and that
limiting the result set to the `max_candidates` size does not help. In
such cases, it may be worth it to determine the size of the group in a
separate `.count()` query, and proceed to select the IDs only if we know
the result set will be small enough.

The included span measurements will capture further performance data,
to ensure that the speed improvement justifies the cost of making the
second query.